PlaneOfBlocks: Fix more integer overflows
Cherry picking dubhater/vapoursynth-mvtools@dfbf72c

+++ src/PlaneOfBlocks.cpp
@@ -1105,7 +1105,7 @@ void PlaneOfBlocks::PseudoEPZSearch()
     sad = LumaSAD(GetRefBlock(0, zeroMVfieldShifted.y));
     sad += saduv;
     bestMV.sad = sad;
-    nMinCost = sad + ((penaltyZero*sad)>>8); // v.1.11.0.2
+    nMinCost = sad + (((int64_t)penaltyZero*sad)>>8); // v.1.11.0.2
 
     VECTOR bestMVMany[8];
     int nMinCostMany[8];
@@ -1124,7 +1124,7 @@ void PlaneOfBlocks::PseudoEPZSearch()
         + SADCHROMA(pSrc[2], nSrcPitch[2], GetRefBlockV(globalMVPredictor.x, globalMVPredictor.y), nRefPitch[2]) : 0;
     sad = LumaSAD(GetRefBlock(globalMVPredictor.x, globalMVPredictor.y));
     sad += saduv;
-    int cost = sad + ((pglobal*sad)>>8);
+    int cost = sad + (((int64_t)pglobal*sad)>>8);
 
     if ( cost  < nMinCost || tryMany)
     {
